updated to build script, check for golang then install

+++ script/build.sh
@@ -8,16 +8,18 @@ sudo apt-get install -y binutils-arm-linux-gnueabi
 sudo apt-get install -y libncurses5-dev
 sudo apt-get install -y gcc-arm-linux-gnueabi
 
-echo "Downloading golang..."
-sudo wget https://dl.google.com/go/go1.10.2.linux-amd64.tar.gz
-sudo tar -C /usr/local -xzf go1.10.2.linux-amd64.tar.gz
-sudo rm go1.10.2.linux-amd64.tar.gz
+if [ ! -d "/usr/local/go" ]; then
+    echo "Downloading golang..."
+    sudo wget https://dl.google.com/go/go1.10.2.linux-amd64.tar.gz
+    sudo tar -C /usr/local -xzf go1.10.2.linux-amd64.tar.gz
+    sudo rm go1.10.2.linux-amd64.tar.gz
 
-echo "Configuring golang..."
-mkdir -p $HOME/go/bin
-sudo echo "GOPATH=$HOME/go" >> /etc/profile
-sudo echo "PATH=/usr/local/go/bin:$GOPATH/bin:$PATH" >> /etc/profile
-source /etc/profile
+    echo "Configuring golang..."
+    mkdir -p $HOME/go/bin
+    sudo echo "GOPATH=$HOME/go" >> /etc/profile
+    sudo echo "PATH=/usr/local/go/bin:$GOPATH/bin:$PATH" >> /etc/profile
+    source /etc/profile
+fi
 
 echo "Getting source code..."
 go get -u github.com/dustinengle/bwk-dash
